+++ devs/docs/es-backports.rst
@@ -432,10 +432,10 @@ should be crossed out as well.
 - [ ] d8b43c62838 Make Snapshot Deletes Less Racy (#54765) (#55226)
 - [x] 156e5aa77f0 Fix testKeepTranslogAfterGlobalCheckpoint (#55868)
 - [ ] e164c9aaee5 Remove Redundant Cluster State during Snapshot INIT + Master Failover (#54420) (#55208)
-- [ ] 48048646e79 Move Snapshot Status Related Method to Appropriate Places (#54558) (#55209)
-- [ ] a610513ec76 Provide repository-level stats for searchable snapshots (#55051)
-- [ ] 52bebec51f6 NodeInfo response should use a collection rather than fields (#54460) (#55132)
-- [ ] 98fba920226 Fail sniff  process if no connections opened (#54934)
+- [x] 48048646e79 Move Snapshot Status Related Method to Appropriate Places (#54558) (#55209)
+- [s] a610513ec76 Provide repository-level stats for searchable snapshots (#55051)
+- [s] 52bebec51f6 NodeInfo response should use a collection rather than fields (#54460) (#55132)
+- [s] 98fba920226 Fail sniff  process if no connections opened (#54934)
 - [x] 65713743c2a Update translog policy before the next safe commit (#54839)
 - [x] 619028c33e0 Implement transport circuit breaking in aggregator (#54927)
 - [s] 475b210eec4 Improve guidance on removing default mappings. (#54915)
